# backend/recap/api/recap/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Avito Recap API
  version: 1.0.0
  description: |
    API MVP «Итоги года» (Avito Recap).

    Сценарий: пользователь выбирает тестовый профиль (`GET /profiles`),
    запускает генерацию итогов (`POST /recaps`), смотрит слайды (`GET /recaps/{id}`)
    и делится публичной карточкой (`POST /recaps/{id}/share`).

    Соглашения:
      * все идентификаторы — UUID v4 (строка, формат `uuid`);
      * все ошибки возвращаются единым телом `Error` с машиночитаемым `code`;
      * даты и время — RFC 3339 в UTC;
      * денежные величины не раскрываются точно: только диапазоны (`AmountRange`);
      * приватность: личный recap и публичная share-карточка имеют разный набор полей.
        В `ShareCard` не попадают названия конкретных объявлений, суммы, переписки
        и любые данные контрагентов.

    Генерация детерминирована: одинаковые входные данные профиля и год
    дают одинаковый recap (те же слайды, тот же архетип).
servers:
  - url: /api/v1
    description: Текущий хост
...
paths:
  /profiles:
    get:
      tags:
        - Profiles
      operationId: listProfiles
      summary: Получить список тестовых профилей
      description: |
        Возвращает набор заранее подготовленных профилей, на которых
        демонстрируется работа генерации. Список небольшой и постраничности не требует.
      parameters:
        - name: year
          in: query
          required: false
          description: |
            Год, для которого проверяется наличие уже сгенерированного recap
            (влияет только на поле `recap` в ответе). По умолчанию — предыдущий год.
          schema:
            $ref: "#/components/schemas/Year"
      responses:
        "200":
          description: Список профилей
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Profile"
              example:
                items:
                  - id: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f01
                    name: Пётр
                    surname: Будаев
                    avatarUrl: https://cdn.example.com/avatars/018f4c1e.png
                    registeredAt: "2019-03-14T10:12:00Z"
                    hint: Активный покупатель техники
                    recap:
                      id: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1fa0
                      year: 2025
                      status: ready
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /recaps:
    post:
      tags:
        - Recaps
      operationId: createRecap
      summary: Сгенерировать «Итоги года» для профиля
      description: |
        Считает метрики профиля за год, определяет архетип и собирает слайды.
        В ответе возвращается только идентификатор — за самими слайдами
        клиент идёт в `GET /recaps/{id}`.

        Операция идемпотентна по паре (`profileId`, `year`):
          * если recap ещё не существует — `201 Created`;
          * если уже существует — `200 OK` с идентификатором ранее сгенерированного recap.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecapRequest"
      responses:
        "200":
          description: Recap для этого профиля и года уже был сгенерирован ранее
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRecapResponse"
        "201":
          description: Recap сгенерирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRecapResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Профиль не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: profile_not_found
                message: Профиль не найден
                requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f99
        "409":
          description: |
            Недостаточно данных за выбранный год, чтобы собрать осмысленный recap
            (например, у профиля нет ни одной активности).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: not_enough_activity
                message: За 2025 год у профиля меньше 5 действий, генерация невозможна
                requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f98
        "422":
          description: Тело запроса синтаксически корректно, но значения недопустимы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: validation_error
                message: "Поле year вне диапазона 2015..2026"
                requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f97
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /recaps/{id}:
    parameters:
      - $ref: "#/components/parameters/RecapId"
    get:
      tags:
        - Recaps
      operationId: getRecap
      summary: Получить «Итоги года»
      responses:
        "200":
          description: Recap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recap"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /recaps/{id}/share:
    parameters:
      - $ref: "#/components/parameters/RecapId"
    post:
      tags:
        - Recaps
      operationId: shareRecap
      summary: Создать публичную share-карточку
      description: |
        Публикует урезанную версию итогов по постоянной ссылке.

        В карточку попадают только безопасные агрегаты: архетип, число активных дней,
        любимая категория, бейджи и «мягкие» метрики. Никогда не попадают:
        суммы сделок, конкретные объявления, сообщения и данные контрагентов.

        Операция идемпотентна: повторный вызов для того же recap возвращает
        ту же ссылку (`200 OK`), если карточка ещё не отозвана.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareRecapRequest"
      responses:
        "200":
          description: Карточка уже была создана ранее
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResponse"
        "201":
          description: Карточка создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Recap ещё не готов — шерить нечего
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: recap_not_ready
                message: Итоги ещё генерируются
                requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f96
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    RecapId:
      name: id
      in: path
      required: true
      description: Идентификатор recap
      schema:
        $ref: "#/components/schemas/Uuid"
  responses:
    BadRequest:
      description: Некорректный запрос (невалидный UUID, битый JSON, неверные типы полей)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: bad_request
            message: Поле id не является UUID
            requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f95
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: recap_not_found
            message: Recap не найден
            requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f94
    TooManyRequests:
      description: Слишком много запросов
      headers:
        Retry-After:
          description: Через сколько секунд можно повторить запрос
          schema:
            type: integer
            format: int32
            minimum: 1
            example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: rate_limited
            message: Слишком много запросов, попробуйте позже
            requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f93
    InternalError:
      description: Внутренняя ошибка сервиса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: Внутренняя ошибка, попробуйте позже
            requestId: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f92
  schemas:
    Uuid:
      type: string
      format: uuid
      example: 018f4c1e-0a5f-7b3a-9c11-2b6a5d0e1f01
    Year:
      type: integer
      format: int32
      minimum: 2015
      maximum: 2026
      example: 2025
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: |
            Пояснение для разработчика: логи, отладка, Sentry.
            Не предназначено для показа пользователю — текст в UI фронт
            подбирает сам по значению `code`.
          example: Профиль не найден
        requestId:
          allOf:
            - $ref: "#/components/schemas/Uuid"
          description: Идентификатор запроса для поиска в логах
    ErrorCode:
      type: string
      description: |
        Машиночитаемый код ошибки:
          * `bad_request` — невалидный JSON, типы или формат идентификатора;
          * `validation_error` — значения полей вне допустимых границ;
          * `profile_not_found` — профиль не существует;
          * `recap_not_found` — recap не существует;
          * `not_enough_activity` — за год недостаточно активности для генерации;
          * `recap_not_ready` — recap ещё в статусе `generating` или `failed`;
          * `rate_limited` — превышен лимит запросов;
          * `internal_error` — необработанная ошибка сервиса.
      enum:
        - bad_request
        - validation_error
        - profile_not_found
        - recap_not_found
        - not_enough_activity
        - recap_not_ready
        - rate_limited
        - internal_error
    Profile:
      type: object
      required:
        - id
        - name
        - surname
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
          example: Пётр
        surname:
          type: string
          example: Будаев
        avatarUrl:
          type: string
          format: uri
          nullable: true
        registeredAt:
          type: string
          format: date-time
          example: "2019-03-14T10:12:00Z"
        hint:
          type: string
          description: Короткое описание профиля для витрины выбора («что за пользователь»)
          example: Активный покупатель техники
        recap:
          allOf:
            - $ref: "#/components/schemas/RecapRef"
          nullable: true
          description: Ссылка на уже сгенерированный recap за запрошенный год, если он есть
    RecapRef:
      type: object
      required:
        - id
        - year
        - status
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        year:
          $ref: "#/components/schemas/Year"
        status:
          $ref: "#/components/schemas/RecapStatus"
    RecapStatus:
      type: string
      enum:
        - generating
        - ready
        - failed
      example: ready
    CreateRecapRequest:
      type: object
      required:
        - profileId
        - year
      properties:
        profileId:
          $ref: "#/components/schemas/Uuid"
        year:
          $ref: "#/components/schemas/Year"
    CreateRecapResponse:
      type: object
      required:
        - id
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
    Recap:
      type: object
      required:
        - id
        - profileId
        - year
        - status
        - archetype
        - slides
        - generatedAt
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        profileId:
          $ref: "#/components/schemas/Uuid"
        year:
          $ref: "#/components/schemas/Year"
        status:
          $ref: "#/components/schemas/RecapStatus"
        archetype:
          $ref: "#/components/schemas/Archetype"
        badges:
          type: array
          description: Все бейджи, полученные пользователем за год
          items:
            $ref: "#/components/schemas/Badge"
        slides:
          type: array
          description: Слайды в порядке показа
          minItems: 1
          items:
            $ref: "#/components/schemas/Slide"
        share:
          allOf:
            - $ref: "#/components/schemas/ShareResponse"
          nullable: true
          description: Данные публичной карточки, если пользователь уже делился итогами
        generatedAt:
          type: string
          format: date-time
          example: "2026-01-09T08:30:00Z"
    Archetype:
      type: object
      description: |
        Тип поведения пользователя. Считается детерминированно из метрик,
        поэтому вместе с ним всегда отдаются `reasons` — прозрачное объяснение,
        почему пользователь получил именно этот архетип.
      required:
        - code
        - title
        - description
        - reasons
      properties:
        code:
          type: string
          enum:
            - treasure_hunter
            - collector
            - dealmaker
            - negotiator
            - minimalist
            - explorer
          example: treasure_hunter
        title:
          type: string
          example: Охотник за находками
        description:
          type: string
          example: Вы просматриваете десятки вариантов, прежде чем выбрать один.
        reasons:
          type: array
          description: Метрики, повлиявшие на выбор архетипа
          minItems: 1
          items:
            $ref: "#/components/schemas/ArchetypeReason"
    ArchetypeReason:
      type: object
      required:
        - metric
        - explanation
      properties:
        metric:
          $ref: "#/components/schemas/MetricCode"
        value:
          type: string
          description: Значение метрики в человекочитаемом виде
          example: "1248 просмотров"
        explanation:
          type: string
          example: Это в 3 раза больше, чем у среднего пользователя площадки
    MetricCode:
      type: string
      enum:
        - active_days
        - views
        - favorites
        - favorite_category
        - purchases
        - sales
        - messages
        - interests_shift
    Badge:
      type: object
      description: Ачивка, выданная по правилам генерации
      required:
        - code
        - title
      properties:
        code:
          type: string
          example: deals_10
        title:
          type: string
          example: Десять сделок
        description:
          type: string
          example: Вы закрыли 10 сделок за год
        iconUrl:
          type: string
          format: uri
          nullable: true
        level:
          type: string
          enum:
            - bronze
            - silver
            - gold
          example: silver
...
    SlideType:
      x-ogen-name: RecapSlideType
      type: string
      enum:
        - intro
        - active_days
        - views
        - favorites
        - favorite_category
        - purchases
        - sales
        - messages
        - interests
        - archetype
        - final
    SlideBase:
      type: object
      required:
        - type
        - title
      properties:
        type:
          $ref: "#/components/schemas/SlideType"
        title:
          type: string
          example: 243 дня на Авито
        subtitle:
          type: string
          example: Вы заходили чаще, чем 8 из 10 пользователей
        cta:
          allOf:
            - $ref: "#/components/schemas/Cta"
          nullable: true
    Slide:
      oneOf:
        - $ref: "#/components/schemas/IntroSlide"
        - $ref: "#/components/schemas/ActiveDaysSlide"
        - $ref: "#/components/schemas/ViewsSlide"
        - $ref: "#/components/schemas/FavoritesSlide"
        - $ref: "#/components/schemas/CategorySlide"
        - $ref: "#/components/schemas/PurchasesSlide"
        - $ref: "#/components/schemas/SalesSlide"
        - $ref: "#/components/schemas/MessagesSlide"
        - $ref: "#/components/schemas/InterestsSlide"
        - $ref: "#/components/schemas/ArchetypeSlide"
        - $ref: "#/components/schemas/FinalSlide"
      discriminator:
        propertyName: type
        mapping:
          intro: "#/components/schemas/IntroSlide"
          active_days: "#/components/schemas/ActiveDaysSlide"
          views: "#/components/schemas/ViewsSlide"
          favorites: "#/components/schemas/FavoritesSlide"
          favorite_category: "#/components/schemas/CategorySlide"
          purchases: "#/components/schemas/PurchasesSlide"
          sales: "#/components/schemas/SalesSlide"
          messages: "#/components/schemas/MessagesSlide"
          interests: "#/components/schemas/InterestsSlide"
          archetype: "#/components/schemas/ArchetypeSlide"
          final: "#/components/schemas/FinalSlide"
    IntroSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          properties:
            year:
              $ref: "#/components/schemas/Year"
    ActiveDaysSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - activeDays
          properties:
            activeDays:
              type: integer
              format: int32
              minimum: 0
              maximum: 366
              example: 243
            percentile:
              type: integer
              format: int32
              minimum: 0
              maximum: 100
              description: Процент пользователей, которых обошёл этот профиль
              example: 82
            busiestMonth:
              type: integer
              format: int32
              minimum: 1
              maximum: 12
              example: 11
    ViewsSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - views
          properties:
            views:
              type: integer
              format: int64
              minimum: 0
              example: 1248
            uniqueListings:
              type: integer
              format: int64
              minimum: 0
              example: 934
            topCategoryTitle:
              type: string
              example: Электроника
    FavoritesSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - favorites
          properties:
            favorites:
              type: integer
              format: int64
              minimum: 0
              example: 57
            oldestFavorite:
              allOf:
                - $ref: "#/components/schemas/ListingRef"
              nullable: true
              description: Объявление, которое дольше всех лежит в избранном — повод вернуться к нему
            stillAvailable:
              type: integer
              format: int32
              minimum: 0
              description: Сколько избранных объявлений всё ещё активны
              example: 12
    CategorySlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - category
          properties:
            category:
              $ref: "#/components/schemas/CategoryRef"
            subcategory:
              allOf:
                - $ref: "#/components/schemas/CategoryRef"
              nullable: true
            share:
              type: integer
              format: int32
              minimum: 0
              maximum: 100
              description: Доля активности в этой категории, %
              example: 41
            recommendations:
              type: array
              description: Свежие объявления из любимой категории
              maxItems: 5
              items:
                $ref: "#/components/schemas/ListingRef"
    CategoryRef:
      type: object
      required:
        - id
        - title
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        title:
          type: string
          example: Электроника
    PurchasesSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - purchases
          properties:
            purchases:
              type: integer
              format: int32
              minimum: 0
              example: 11
            topCategoryTitle:
              type: string
              example: Хобби и отдых
            badge:
              allOf:
                - $ref: "#/components/schemas/Badge"
              nullable: true
    SalesSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - sales
          properties:
            sales:
              type: integer
              format: int32
              minimum: 0
              example: 8
            amountRange:
              allOf:
                - $ref: "#/components/schemas/AmountRange"
              nullable: true
              description: Приблизительная сумма продаж — только диапазон
            fastestDealDays:
              type: integer
              format: int32
              minimum: 0
              description: За сколько дней ушло самое быстрое объявление
              example: 1
            badge:
              allOf:
                - $ref: "#/components/schemas/Badge"
              nullable: true
    MessagesSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - messages
          properties:
            messages:
              type: integer
              format: int64
              minimum: 0
              description: Количество сообщений. Содержимое переписок не раскрывается никогда
              example: 342
            chats:
              type: integer
              format: int32
              minimum: 0
              example: 76
            asBuyerShare:
              type: integer
              format: int32
              minimum: 0
              maximum: 100
              description: Доля сообщений, отправленных в роли покупателя, %
              example: 64
    InterestsSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - periods
          properties:
            periods:
              type: array
              description: Как менялись интересы в течение года
              minItems: 1
              items:
                $ref: "#/components/schemas/PeriodInterest"
            shiftSummary:
              type: string
              example: Зимой смотрели технику, а к лету переключились на велосипеды
...
    ArchetypeSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          required:
            - archetype
          properties:
            archetype:
              $ref: "#/components/schemas/Archetype"
    FinalSlide:
      allOf:
        - $ref: "#/components/schemas/SlideBase"
        - type: object
          properties:
            badges:
              type: array
              items:
                $ref: "#/components/schemas/Badge"
            actions:
              type: array
              description: Что предлагается сделать после просмотра итогов
              items:
                $ref: "#/components/schemas/Cta"
    ShareRecapRequest:
      type: object
      properties:
        includeSlides:
          type: array
          description: |
            Какие слайды вынести в публичную карточку. Если не передано —
            берётся безопасный набор по умолчанию (`active_days`,
            `favorite_category`, `archetype`). Слайды с чувствительными данными
            (`sales`, `purchases`, `messages`, `favorites`) в карточку не попадают,
            даже если запрошены.
          items:
            $ref: "#/components/schemas/SlideType"
        theme:
          type: string
          enum:
            - light
            - dark
          default: dark
    ShareResponse:
      type: object
      required:
        - token
        - url
        - card
        - createdAt
      properties:
        token:
          type: string
          description: Публичный токен карточки (не UUID, чтобы id recap не утекал в ссылку)
          example: QmZ4x9Lp
        url:
          type: string
          format: uri
          example: https://recap.example.com/s/QmZ4x9Lp
        imageUrl:
          type: string
          format: uri
          nullable: true
          description: Пререндеренная картинка карточки для соцсетей
        card:
          $ref: "#/components/schemas/ShareCard"
        createdAt:
          type: string
          format: date-time
          example: "2026-01-09T08:35:00Z"
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: "`null` — ссылка бессрочная"
    ShareCard:
      type: object
      description: |
        Публично видимая часть итогов. Содержит только агрегаты, которые
        безопасно показывать посторонним: без сумм, объявлений и переписок.
      required:
        - year
        - archetypeTitle
      properties:
        year:
          $ref: "#/components/schemas/Year"
        displayName:
          type: string
          description: Имя без фамилии
          example: Пётр
        archetypeTitle:
          type: string
          example: Охотник за находками
        archetypeDescription:
          type: string
          example: Вы просматриваете десятки вариантов, прежде чем выбрать один.
        activeDays:
          type: integer
          format: int32
          minimum: 0
          maximum: 366
          example: 243
        favoriteCategoryTitle:
          type: string
          example: Электроника
        badges:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/Badge"
        theme:
          type: string
          enum:
            - light
            - dark
          example: dark